pyproject.toml
[project]
name = "xkcd-archiver"
version = "3.0.0"
description = "Downloads every XKCD comic"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "David Antonini", email = "toonarmycaptain@hotmail.com" },
]
dependencies = [
    "requests>=2.32",
    "textual>=0.50",
    "Pillow>=10.0",
]

[project.scripts]
xkcd-archiver = "XKCD_archiver.downloadXKCD:cli_run"
xkcd-tui = "XKCD_archiver.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["XKCD_archiver"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: tests that hit real xkcd.com API (deselect with '-m not integration')"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
